VIKOR in its own file now
